2315: extract RunButton as a separate component
